chore: add project configuration files
