+++ MANIFEST.in
@@ -1,0 +1,4 @@
+include README.md
+include LICENSE
+include requirements.txt
+recursive-include mouse_patrol *.py
