+++ pyproject.toml
@@ -1,0 +1,48 @@
+[build-system]
+requires = ["setuptools>=61.0", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "mouse-patrol"
+version = "0.2.1"
+description = "Python utility to automatically & regularly move your mouse to prevent going idle"
+readme = "README.md"
+requires-python = ">=3.8"
+license = {file = "LICENSE"}
+authors = [
+    {name = "Leon Bojanowski", email = "leongaborbojanowski04@gmail.com"}
+]
+maintainers = [
+    {name = "Leon Bojanowski", email = "leongaborbojanowski04@gmail.com"}
+]
+keywords = ["mouse", "automation", "idle", "screensaver", "pyautogui"]
+classifiers = [
+    "Development Status :: 3 - Alpha",
+    "Intended Audience :: End Users/Desktop",
+    "License :: OSI Approved :: MIT License",
+    "Operating System :: OS Independent",
+    "Programming Language :: Python :: 3",
+    "Programming Language :: Python :: 3.8",
+    "Programming Language :: Python :: 3.9",
+    "Programming Language :: Python :: 3.10",
+    "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
+    "Topic :: Desktop Environment",
+    "Topic :: System :: Systems Administration",
+    "Topic :: Utilities"
+]
+dependencies = [
+    "pyautogui"
+]
+
+[project.urls]
+Homepage = "https://github.com/f1nal04/mouse-patrol"
+Repository = "https://github.com/f1nal04/mouse-patrol.git"
+Issues = "https://github.com/f1nal04/mouse-patrol/issues"
+
+[project.scripts]
+mouse-patrol = "mouse_patrol.main:main"
+
+[tool.setuptools.packages.find]
+where = ["."]
+include = ["mouse_patrol*"]
